+++ CHANGELOG.md
@@ -3,6 +3,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### ✅ Added
+- Highlighting and tapping on user mentions
+
 ### 🔄 Changed
 
 # [4.52.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.52.0)