+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.53.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.53.0)
+_April 30, 2024_
+
 ### ✅ Added
 - Highlighting and tapping on user mentions
 - Customization of the channel loading view